Remove NotesMigration#enabled()
The name was too vague to be useful. Many callers assumed it meant reads
were enabled, when in fact it was true when writes were enabled as well.
This caused several tests depending on NoteDb-only features, like
CC-by-email, to fail when GERRIT_NOTEDB=write.

The main non-test usage that was somewhat ambiguous was
AbstractChangeNotes#loadRevision, but it turns out this is only used by
ChangeResource#prepareETag, so using readChanges instead is correct.

Change-Id: I83e3732eb3fc0743860d1fd030526f54a611dc7c

Former-commit-id: e8bad33657d840a11c27a47cb3b840821ff1b2e0