Change gr-styled-table to be css template
Previously, gr-styled-table was include in the list view only, which
handled styling the content passed to the list view (which were always
tables of varying content).

Because Polymer2 does not allow for reaching into <slot> to style
content as was done before, the styles have to be imported to the
element that originally creates the content.

Change-Id: I7136f8ccbea00e532d2a8e1a63f6874eb418906e

Former-commit-id: c9ec9a16fbc713d7ad5bef138eca809bbe1ae64f