Fix back button behavior for hash redirects
A URL that starts with a hash is of the old GWTUI format, and should
redirect to the new PolyGerrit style without the hash. The current
implementation called page(newUrl), which adds another entry to the
browser history. Using page.redirect(newUrl) instead modifies the
history in place, enabling the expected back button behavior.

Bug: Issue 6693
Change-Id: I9e0503808332528c7b86ff24ddeb89d56dcf4c09

Former-commit-id: 63d7e35f102977f4f74f82c4c6d5e7fe42e0d67e