Cache negative results in PatchListCache
Computing PatchLists can fail with LargeObjectException. In this case,
computing the PatchList was very expensive in the first place. This
commit adds logic to cache such negative results and don't attempt to
recompute the PatchList since it would yield the same result.

At Google, we see a lot of CPU time being spent on computing PatchLists
for large objects that eventually fail because of LargeObjectExceptions.
Caching negative results will at least prevent pointless retries.

If the LargeObject threshold is changed, host admins can flush the cache
or manually remove the tombstones.

Change-Id: Ia87cebede79fb1c933eb33072ec9a76768938afd

Former-commit-id: 8d26b686cf79e395ec1f718d896f380834b455fd