Recognize all parts marked with 'multipart/'
Gerrit currently only recognizes a subset of multipart MIME messages.
Apache James supports all known types and this change makes it so that
Gerrit does too.

This fixes a bug where inbound emails would not be parsed, because the
message was signed and is therefore wrapped in a 'multipart/signed'
part.

Change-Id: I0de8820188e5295731d9253c9bb3f452da96908a

Former-commit-id: 62f34ab35e5d2f216fbbd72144986993dd25a66c