Revert "Update iron-input to 2.0"
This reverts commit d13b70bd07f377932e3549af5e4c640b40e581cb [formerly 37636a62564b09df8d8e4e48828b72afdf817bb0].

Reason for revert: The googlesource.com environment is not ready for
this yet.

I didn't realize that we couldn't yet use the hybrid version of the
elements in google3 yet. They exist in the polymer2 directory, but
apparently that depends on using polymer2. This change will need to be
reverted until iron-input v1 is updated to get the polymer2 "hybrid"
version.

Change-Id: Ibeeae2458337b0a225993e12b043b1e65c3c4c04

Former-commit-id: 10de10381565fc0f42c0975c9678769ec1637976