Squash comments in plaintext inbound email
In HTML, we have been squashing comments that belong to the same entity
for quite some time now. This works well and this commit expands that
model to plaintext email.

Right now, Gerrit would only take the first comment (e.g. a change
message). If users would then type more text inline to comment on
another user's change message, those comments would be lost. This change
makes it so that those comments are appended to the change message.

Change-Id: Ieb814893262aa6d9d2cd27dd885817cd61ca7dad

Former-commit-id: 37834bc63f78630bda8c6ba56ce905961d505303