Revert "Remove styles for dueToRebase chunks"
This reverts commit 857490a95a6e152bc673701ab17029c984fd1c69 [formerly e048852365d1c7fc92bea2f597f117c13140f9a0].

Reason for revert: Highlighting which hunks were introduced by a rebase
is done as best-effort and will be clearly stated as such when we
announce this feature. Hence, we expect at least some false-negatives
at the moment (which we intend to get rid off over time). There aren't
any false-positives which we know of at the moment.

Change-Id: I103dbe8bea98d0bb7a417c02d371c2864dca015d

Former-commit-id: 4cb04e13cd8683bede567cff63153d4d78fdbf26