Fix list view so that it is not loaded twice
The list view listens for filterChanged, so in addition to the initial
load, upon element initialization, it was getting called, even when
the new & old values were falsy. Add a check that does not trigger a
reload when the filter changes between two falsy values.

Also fixes a small css issue in which the reload text had a grey
background.

Bug: Issue 6816
Change-Id: I301bea7b41adde58d875e6cac20290cb3d6b3ade

Former-commit-id: 9f2f3a68f1174b3c94f3cc5031a380cb30795f65