Include change detail query params in ETags cache key
The diff view requests a subset of the change detail compared with the
change view, but the query params which specify the subset were not
included in the ETags cache key, the cache keys collided and the
incorrect ETag was used in the superset request.

Fixed by including the query params in the key.

Also fix spelling in a method name.

Bug: Issue 6818
Change-Id: If5f17451d5a5dc13a7c315d518bf47b595546b9d

Former-commit-id: 47e27a8b9a5e809b307059dd63e0b9ff66b842b9