Allow to push updates of account.config file
Pushing updates of account.config files was disabled while accounts were
stored in both ReviewDb and NoteDb to prevent that the accounts go out
of sync between ReviewDb and NoteDb. Now since accounts only live in
NoteDb we can allow pushing updates of account.config files. However
pushing updates to an account.config file is still rejected if:

- it cannot be parsed
- the preferred email is invalid
- the own account is deactivated

The same checks are done on submit if updates of account.config files
have been pushed for review.

Pushing an account.config file with an invalid preferred email is not
rejected if the preferred email didn't change. This means if an account
already had an invalid preferred email updates of account.config can
still be pushed without being forced to fix the preferred email. However
if the preferred email is changed it is verified that is valid.

When validating the account.config file on submit we don't have the old
file version available and hence we can't check if the preferred email
has changed. This is why in this case we always enforce the preferred
email to be valid.

Change-Id: Ie08e50ea92fa7101c5cfcdca78632e8fd52cb26b
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: ed7c347ddb262ce9faa7edd879b057005ea22045