Handle more autocomplete cases
When the tab-complete property of gr-autocomplete is true, the user
expects tab to complete to the current selected autocomplete entry and
then retain focus on the input for more typing.

Previously, due to gr-autocomplete-dropdown key handling, focus never
returned to the input. This change fixes that and adds more test cases
to prevent further regressions.

Bug: Issue 6253
Change-Id: I63943e2400e4f7f61fcc4437fd5ed317066843d7

Former-commit-id: 62bae3e056a2f8cd5646a740abc7ccd5a6e2f1e8