Add function upgradeUrl to gr-navigation
The migration to project-based URLs has made evident the need for some
alias in gr-navigation to `history.replaceState`. This change adds
the function `upgradeUrl` to the gr-navigation abstraction to handle
this use case.

Any external project that does not wish to support this functionality
should implement this function as a no-op, e.g. () => {}.

Next steps:
1 In gr-diff-view and gr-change-view, add 'project' param to app.params
  when getChangeDetail returns a response.
2 In gr-diff-view and gr-change-view, call upgradeUrl in getChangeDetail
  iff (!app.params.project || project !== app.params.project).
3 Modify generateUrl function to generate project based URLs.

Bug: Issue 6708
Change-Id: Ie9bea7f9766e784fa30e5692582500bab3f1642f

Former-commit-id: de8d2ff679b3b41cbf0753741f200b3f38264a23