Always post a message when starting review
This is a temporary workaround for issue 6841. PostReview does not send
any notifications if there are no comments published and no message.

By detecting this case and inserting a dummy message, we can force an
email to be sent. Once issue 6841 is fixed on the server side, we can
revert this change.

Change-Id: Ie2963546b4d9bc7b37f52c3f47110a05058407ac

Former-commit-id: 3f1a6ff4563c8c3f6107c0a72d46ca9d20302149