AbstractDaemonTest: Toggle draft status of change and patch set
This is a preparation change for disabling creation of draft changes. We
still want to be able to test publishing, deleting and abandoning of
draft changes and patch sets. Move the operation for setting draft status
for change and patch set from DraftChangeIT to AbstractDaemonTest. Add a
test to change the draft status and patch set status to draft even when
the change.allowDrafts is disabled, by pushing to refs/for/master and
toggling draft status.

Change-Id: I1148e28b5364b7d9d635909efa19a089cb7072c3

Former-commit-id: 1974b6892a5ef94990aeab30cc89b05c671d0a8f