Return project in GetRelated
This commit adds project as a mandatory return field in
RelatedChangeAndCommitInfo returned by GetRelated to fix a bug where
links to related changes would have 'undefined' as the project component.

Since we now always want to have the project along the change number it
seems to be a good practice to also return the project along with a
change number in the API similar to ChangeInfo.

Bug: Issue 6876
Change-Id: I64ac0d1b30a3b8034cf16afe96c9d1168cbde193

Former-commit-id: cab63519a68d5fb53bd7ad2c7877865353ed37b9