JsPlugin: Add handling for .html files
Change 61a82af3b7 [formerly 2b2fdb749b] added support for files with .html extension to be
loaded as javascript plugins, but did not add corresponding support
in JsPlugin which still assumes that all javascript plugins will
have .js extension.

As a result, attempting to get index of the ".js" extension on a file
with ".html" extension results in StringIndexOutOfBoundsException.

Bug: Issue 6890
Change-Id: I2b0e14387fac5907617ca8cc551997c3d7192033

Former-commit-id: 78af89ba90aac3ab9ab410480d3cccc492a9c4a4