Allow external IDs for account with ID 0
Account IDs are expected to be positive but somehow we have a least one
account with ID 0. External IDs for this account are ignored when the
external IDs are loaded and an error is logged. Relax the account ID
check when external IDs are loaded and allow external IDs for accounts
with ID 0. On writing external IDs for such an account there is no
problem, since any account ID that resolves to an existing account is
allowed.

Change-Id: I50091a54ebb803110fe0332f60b49615ffb9df34
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 8e27528ddc11ce57c86ea48fc34d2660c4dd54f2