Implement _projectLookup in gr-rest-api-interface
This change creates a new object, _projectLookup, for use as a shortcut
in supporting slicer.

It is a map of changeNums to projects, with added logic to fetch the
change and populate the lookup if it has not already been populated.

This object is then populated with data in the router and rest API.

Bug: Issue 6708
Change-Id: Ide5bf94539d7e97c929d17505deb59aaf0e6c192

Former-commit-id: b056fb5584a41d986f0e3471918ef917fc878f60