Fix OR index queries in ChangeIndexRewriter
If a query was like "A OR B" where both A and B are subexpressions
that can be answered by the index, the QueryRewriter was creating a
predicate tree like:

  (status:new OR status:draft)
  AND
    A OR B

This was caused by the two children A, B being rewritten but then
wrapped in a standard OrPredicate instead of an OrSource.  With the
root of the predicate tree not being a ChangeDataSource the rewriter
added effectively "(is:open (A OR B))" to ensure the query could
execute.

A simple fix is to check if all rewritten children are ChangeDataSource,
and if so, wrap them with the proper AndChangeSource or OrSource
predicate implementation.

Change-Id: Ia57fc59edcdf9d9edce2164f55d2d5b74e32fb36

Former-commit-id: 48373ee948023894603aae9675817c8b4cc36381