Move ReceivePack creation into AsyncReceiveCommits
The general idea of the AsyncReceiveCommits->ReceiveCommits delegation
is that ReceiveCommits does all its work in a background thread whose
lifecycle is managed by AsyncReceiveCommits. Doing significant work in
the ReceiveCommits constructor to build a ReceivePack doesn't really fit
with that, especially considering that AsyncReceiveCommits then
immediately pulls the ReceivePack out of the ReceiveCommits and holds a
reference to it.

There was previously a reason that ReceiveCommits wanted to directly
manipulate the ReceivePack setters, which was to populate the allRefs
field. This logic has now been moved out to AllRefsWatcher, so at this
point there was no remaining reference from the ReceivePack to the
containing ReceiveCommits. It is now easy to move ReceivePack
construction into AsyncReceiveCommits.

This will make it possible (albeit still nontrivial) to make the entire
ReceiveCommits operation retryable.

Change-Id: I41bc8221bac01a0a88fc188b7a7fc3367a3f0267

Former-commit-id: d9c7684fcb2ddb880b497705b20755e186f7e180