Add new topic, Changes, to documentation
This is a new concept topic for the Gerrit documentation. Note that,
while this topic contains a collection of content from a variety of
other topics in the doc set, such as the Review UI topic, the User
Guide topic, and the Project Owner Guide. Those topics will remain,
as they are still useful to users.

This topic is in many ways a reference that future "how-to" topics can
point to. The intended audience is Gerrit users who want more in-depth
information about what a change is, and what information about a change
they can view from the Gerrit UI.

Note: We may add screenshots to this topic in the future.

Change-Id: Iea3ba2aa7cdc77b561cb95f0ea83b1a3193bc281

Former-commit-id: 6493b2f25a0d6a1f283eae77a52f46c841525213