AccountManager: Improve error message on first user creation failure
It appears to be particularly hard to troubleshoot the failures on
first user creation and why it wasn't granted Administrator privileges.
It seems to be a common pattern that the sites are initialized on non
empty All-projects repository, in which case, the group UUID in the
database and in the All-Projects repository don't match. If this
happens, NSEE is reported:

java.util.NoSuchElementException
   at java.util.ArrayList$Itr.next(ArrayList.java:854)
   at com.google.gerrit.server.account.AccountManager.create(AccountManager.java:255)
   at com.google.gerrit.server.account.AccountManager.authenticate(AccountManager.java:118)
   at com.google.gerrit.httpd.auth.ldap.LdapLoginServlet.doPost(LdapLoginServlet.java:122)
   at javax.servlet.http.HttpServlet.service(HttpServlet.java:648)
   at javax.servlet.http.HttpServlet.service(HttpServlet.java:729)

Check for this specific error pattern and report a more specific error
message.

Change-Id: I2157b7341c358d0db1726a72a02d1bd574e55284

Former-commit-id: a701936040635c2aa5bef48f7780e02c01272680