Modify quoted comment truncation rules
When a diff comment replies to another, the resulting email message
includes a quotation of the parent message with some truncation.

Formerly, messages were truncated to 75 characters or the first line of
the comment (whichever was shorter). The resulting truncated quotation
presented no indication that it had been truncated.

With this change, the message is truncated to 100 characters or the
first line of the comment, or the after the last period within 100
characters (whichever is shorter). If the message is truncated as a
result of these rules, an ellipsis is appended.

Bug: Issue 6111
Change-Id: I0e9ffb8ebd7c734860021d7a6948b181daccb034

Former-commit-id: ac5797400f199cac3baaa0808f84443c738667b3