Avoid repeating scans of refs/{heads,tags} in getAlreadyAccepted
FluentIterable can be run multiple times, each time the iterator is
accessed per change considered in the ConflictsPredicate.  That can be
expensive on the repository.  Instead use Java 8 streams and convert
to a Set<ObjectId> immeidately.

Fix the addCommits method to also consider annotated tags that point
to a commit. It was adding lightweight tags, but not annotated tags
as it failed to peel the RevTag object and can't use parseCommit()
(which auto-peels), because it doesn't want to catch an exception if
a ref points at a non-commit.

Change-Id: Ifc8285de88fa04de60d4726303ec7987f256fcd1

Former-commit-id: 942eea21177289e5ce8ac9f362cdf22dc662abb3