RetryHelper: Refactor to take multiple options
Pass an AutoValue Options arg in instead of one (and soon more than one)
nullable argument. The Options arg is a slightly more user-friendly
wrapper around upstream's RetryerHelper.

Use this approach instead of making RetryHelper a non-singleton and
using a more builder-ish approach for RetryHelper itself. This would
require much more churn at call sites, the vast majority of which are
fine with the defaults.

Change-Id: I285843940c06d9e520b611d9bfa5053634682ce1

Former-commit-id: c9d84f74049e80279e1bb33f32afb546db446431