MergeOp: Multiply retry timeout by number of projects
RetryHelper uses a single Retryer for the entire operation, which in the
case of MergeOp may potentially include updating a large number of
repositories independently. This means that the first lock failure
encountered during MergeOp may have happened after eating up most or all
of the default 20s retry budget.

Work around this by increasing the retry timeout based on the total
number of projects in the ChangeSet.

Change-Id: Ie8f078d5691de56fd2bfc2946358f4a1f0930a06

Former-commit-id: 3575a3771f3a91227f0f12bc4afed4899253114e