Start moving index code to gerrit-index directory
Start by renaming the existing gerrit-antlr package to gerrit-index.
This naming reflects what the package is used for, not the
implementation detail that ANTLR is used for the parser.

Also rename the server.query package to index.query, to match the
top-level directory name, rather than mixing into the server package
"owned" by gerrit-server.

Move QueryParserTest into gerrit-index, next to the code it's testing.

Change-Id: I629fe4c7d15c76767a190ccf180772b8d7c96984

Former-commit-id: 62a434ec706ba42fff5f44e031644f441be990a3