Always setLastLoginExternalIdKey for IdentifiedUsers
Anonymous users have no login external id key, so it makes no sense to
set it for them in the read path.

The two callers for this property are GetExternalIds and
DeleteExternalIds.  The latter is not a read method and should expect
to have the current value to prevent the user from deleting their
current external id.

Change-Id: Id0671530fe649e6b0fcfd489fc384e37a2fd02e2

Former-commit-id: 7cf423d6088ac9656fc240a904e4197dc88478cc