Make file list URLs explicitly depend on patchNum
The Polymer binding to generate diff URLs had depended on the patch
range object, but this would be created before the patch number was
known, leaving it undefined inside the range object. With this change,
the `_computeDiffURL` method explicitly depends on the range components,
rather than the object that houses them.

Bug: Issue 6978
Change-Id: I871a9f340057cc7db8eda143d7bf110ea955019b

Former-commit-id: acc46838db8f7b20f0b29afb5cb6113378e3e35e