Record RestModifyView name in NoteDb reflog
Instead of using a generic string "Update NoteDb refs", record the
class name of the RestModifyView that caused this operation to
execute.  This makes inspection of the history of a change much
easier, as its possible to correlate events to specific sections of
the Gerrit server.

The implementation favors going from the RestApiServlet down the
stack, so that plugin handlers are caught first and logged as the
plugin's RestModifyView implementation, even if the plugin is using
an internal API call to a standard view like PutTopic.

Change-Id: Ifd6a9843b6bd5f3e7f1ff7e145fa433f237b147a

Former-commit-id: 6f42e5d2d58a5031f3f9f63419b2ea4789aaf9f5