Stub inline render function
The 'n key presses' suite of gr-file-list tests would trigger async
inline diff rendering by an effect. Because the async rendering is
niether explicitly started nor cancelled, it extends past the lifetime
of the test and render work goes on in the background of the following
test. The processing code called functions which were stubbed when
triggered but were un-stubbed at the end of the test (sometimes before
the async processing had completed). The processing attempts to invoke
such a function the failure is attributed to whichever test is running.

With this change, the async rendering is not triggered in the first
place by setting an additional stub.

Bug: Issue 6949
Change-Id: I06dd2e0c889937a353da30337215c813db32ef42

Former-commit-id: 40b74b6f5d056d1e4212222b62cf6a2d0c784088