Modify gr-navigation schema to match app.params
_generateUrl in gr-router expected project names to be passed in with
the 'project' property on the params object parameter. However, utility
functions in gr-navigation passed it with the prop name 'projectName' --
this meant that no generated URLs would ever have the project name
supplied.

This change fixes this inconsistency, and modifies the getUrlForChange
function to also supply the project name to generateUrl.

Change-Id: I05c1ef21db16278e0795aee3fb6dbf76601ad6f5

Former-commit-id: 4cd9917e38902742933999779ec202b597307eea