PolyGerrit: gr-main-header -> Rename "More"-button to "Browse"
The "More"-button brings the user to a screen where projects can
be listed. From this screen the user can also search and browse
available groups.

Going from the old gwt-ui, it is certainly not easy to find the
project/group listing screen this way.

By renaming "More" to "Browse" the user will be at least hinted
about which features are provided by the link destination.

Change-Id: I2e72809c145e3bb6a0d3f3944b45da709a72c209

Former-commit-id: 33c9f552b5a8fd2af50e726e9a5cec6cb70ec0de