Add support for viewing other users' dashboards
The PolyGerrit router already accepts URLs like /dashboard/USER, but we
didn't respect the parameter. This change replaces all occurrences of
"self" that were hard-coded in the dashboard section queries with a
${user} placeholder.

We also eliminate redundant set of search query requests that are
emitted when navigating away from the dashboard view.

Feature: Issue 6982
Change-Id: I7baad12a02e0e9e53b74537215ddbf21c7b2d006

Former-commit-id: 9ccf9d023bc486d015a02c019ba6672a7b0b831b