Introduce gr-rule-editor
This is the most granular component in the project access editor.
It allows the user to modify the value of a particular rule.

Here is the access hierarchy:

Project access
- Sections
  - Permissions
    - Rules <-- This is the component for this change.

It will be fed in the group, permission, optional labels, ruls, and
section from its parent(s).

There are a variety of different ways this can behave that are outlined
in the tests.  Generally they are:
- Some permission types have different option values
- When there are label options, they should be presented
- Some permission types should display a push option, and their labels
  vary based on whether it is an edit or a push.

Change-Id: Iaf0aab09418c4e2589753419d82c375d8d5cf673

Former-commit-id: 7198afd8ef0e60e20171173300118363a21a674e