Move gerrit-server/prolog sources to top-level prolog directory
The current layout of having gerrit-* directories with separate java
source trees dates back to the bad old days of Maven, when this specific
layout was necessary to decompose the build into separate pom.xml files.
This is no longer a requiement.  This series rearranges the directory
structure and dissoves all top level directories and move all sources to
the programming languages top level directories and differentiates
between sources and tests.

So that the gerrit-* will be dissolved and moved to new top level source
directories:

* java
* javatests
* prolog
* resources

Lib, polygerrit-ui and tools top-level directories stay the same.

See this discussion on the mailing list:
https://groups.google.com/d/topic/repo-discuss/TcvEkD-KO_c/discussion

Change-Id: I3aff3a9787189cbfc79ec95537a32bed1584e444

Former-commit-id: 71f54a5508dfbfbd261d13f5bf87f264e795cc09