Rebase: Use BooleanCondition for #setVisible
The UIAction in Rebase is a special case, because it contains an
expensive operation (open repository) which we only performed if all
prerequisites for a rebase were matched.

This commit adapts the implementation to find a balance between
performance and calls to PermissionBackend. The permission is now added
to the conditions at the end of the method. This means we perform
opening the repository more frequently than we used to, but spare the
individal PermissionBackend call by adding it to the batch.

For Rebase, this seems fine as most users will have the rebase
permission if they can see the change. An example is gerrit.gs.com where
users can upload new patches and perform rebases on changes they don't
own.

Change-Id: I4a6316df3a3b6f50ed41c45b3fbfb8e3835dbd7a

Former-commit-id: becae7acd36d4e1c223719be270061d183f01b83