Store metaId in account
The account index should contain fields that allow us to check if an
account in the index is stale. For this we need a state ID for each
source that contributes to the account (user branch + external IDs).
For the user branch we can record its ref state (similar to how the
change index records ref states / ref state patterns for all refs that
contribute to a change). For this we store the SHA1 of the user branch
in the account when it is loaded or updated. The metaId is only
available if reading accounts from NoteDb is enabled.

The metaId inside Account is stored as String, since the Account class
is also used on client-side and can't depend on JGit's ObjectId type.

For the external IDs we will record the IDs of the note blobs in the
external IDs branch. Making these blob IDs available is implemented in
change I53e5d9a9a6.

A follow-up change will add the new fields to the account index that
store the ref state of the user branch and the blob IDs of the external
IDs.

Change-Id: Ideaead9b1d875b2b35f52004aca4854403e067f8
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 1601a115bd6c755c428994cf1c0ea704b016bcbd